Switched pwp to long double

+++ cPWP/bamsToPWP.cpp
@@ -14,7 +14,7 @@
 #include <sstream>
 #include <string>
 #include <vector>
-//#include <algorithm>
+#include <algorithm>
 #include <thread>
 #include <boost/iostreams/filtering_stream.hpp>
 #include <boost/iostreams/filtering_streambuf.hpp>
@@ -226,8 +226,10 @@ int calcPWPfromBinaryFile (std::string binaryFile, unsigned long long int numLoc
          First, we'll generate all of these vectors, which apparently in C++ needs to be constructed of a 
          vector of two-dimensional vectors... 
          */
-        std::vector<std::vector<std::vector<unsigned long long int>>> pwpThreads(numThreads, std::vector<std::vector<unsigned long long int>> (numIndividuals, std::vector<unsigned long long int> (numIndividuals,0) ) ); //pwpThreads[0] is the first 2D array for the first thread, etc...
+        std::vector<std::vector<std::vector<long double>>> pwpThreads(numThreads, std::vector<std::vector<long double>> (numIndividuals, std::vector<long double> (numIndividuals,0) ) ); //pwpThreads[0] is the first 2D array for the first thread, etc...
         std::vector<std::vector<std::vector<unsigned long long int>>> weightingsThreads(numThreads, std::vector<std::vector<unsigned long long int> > (numIndividuals, std::vector<unsigned long long int> (numIndividuals,0) ) );
+        
+
 
         // Now we need to determine how many loci for each thread. If we want to use the entire binary file, instead of numLoci loci, then change this to lociPerThread = (size/(numIndividuals*2))/numThreads
         //unsigned long long int lociPerThread = numLoci / numThreads;
